% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{a}
\alias{a}
\title{Generate HTML tag for element a}
\usage{
a(
  ...,
  accesskey = NULL,
  autocapitalize = NULL,
  class = NULL,
  contenteditable = NULL,
  dir = NULL,
  download = NULL,
  draggable = NULL,
  hidden = NULL,
  href = NULL,
  hreflang = NULL,
  id = NULL,
  inputmode = NULL,
  is = NULL,
  itemid = NULL,
  itemprop = NULL,
  itemref = NULL,
  itemscope = NULL,
  itemtype = NULL,
  lang = NULL,
  part = NULL,
  ping = NULL,
  referrerpolicy = NULL,
  rel = NULL,
  slot = NULL,
  spellcheck = NULL,
  style = NULL,
  tabindex = NULL,
  target = NULL,
  title = NULL,
  type = NULL,
  custom_attr = NULL
)
}
\arguments{
\item{...}{A string or strings of permitted content for this HTML element tag. The user is responsible for determining what is permissible.}

\item{accesskey}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{autocapitalize}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{class}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{contenteditable}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{dir}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{download}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{draggable}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{hidden}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{href}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{hreflang}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{id}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{inputmode}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{is}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemid}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemprop}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemref}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemscope}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemtype}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{lang}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{part}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{ping}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{referrerpolicy}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{rel}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{slot}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{spellcheck}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{style}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{tabindex}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{target}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{title}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{type}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{custom_attr}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}
}
\value{
A HTML tag string.
}
\description{
Generate HTML tag for element a
}
\examples{
a(class = "test", "Example")
}
