% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{abbr}
\alias{abbr}
\title{Generate the <abbr> HTML tag.}
\usage{
abbr(..., attr = NULL, separate = FALSE, collapse = "")
}
\arguments{
\item{...}{A string or vector of content to pass to the tag.}

\item{attr}{A named list or named vector, names are attribute names and values are attribute values.}

\item{separate}{TRUE/FALSE, If TRUE, returns separate tags for each item of ...; if FALSE, returns one tag with the items of ... in the tag content.}

\item{collapse}{A string. If NULL, returns a vector the same length as ... instead of collapsing the tags into one string.}
}
\value{
A HTML tag string.
}
\description{
The <abbr> HTML element represents an abbreviation or acronym; the optional title attribute can provide an expansion or description for the abbreviation. If present, title must contain this full description and nothing else.
}
\details{
Learn more at \url{https://developer.mozilla.org/en-US/docs/Web/HTML/Element/abbr}.
}
\examples{
abbr(attr = list(class = "test"))
}
