% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{attr_helper}
\alias{attr_helper}
\title{Helper function to generate HTML5 attribute strings}
\usage{
attr_helper(attr, separate = FALSE)
}
\arguments{
\item{attr}{A named list, names are attribute names and values are attribute values.
If the items of the list and the items of the tag content are longer than length 1,
the items for the attribute will correspond with the items of the content in the same position.
(ex. when generating a series of option tags, you might want to pass a different id attribute for each
item of the content, you can pass the vector of ids in the named list of attributes)}

\item{separate}{TRUE/FALSE, if TRUE, returns a vector for creating multiple tags at once.}
}
\value{
A HTML attribute string.
}
\description{
Helper function to generate HTML5 attribute strings
}
\examples{
attr_helper(attr = list(class = 'test'))
}
