% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\docType{data}
\name{boolean_attributes}
\alias{boolean_attributes}
\title{Vector of boolean attributes}
\format{
A vector
}
\usage{
boolean_attributes
}
\description{
A vector of boolean attributes
}
\keyword{datasets}
