% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{content_indenter}
\alias{content_indenter}
\title{Add new lines and tabs to format HTML content.}
\usage{
content_indenter(x)
}
\arguments{
\item{x}{A string of HTML to format.}
}
\value{
A HTML string formatted with new lines and tabs.
}
\description{
Add new lines and tabs to format HTML content.
}
