% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{doctype}
\alias{doctype}
\title{Generate HTML document string with properly declared DOCTYPE.}
\usage{
doctype(..., doctype = "html")
}
\arguments{
\item{...}{A string or strings of HTML element tags.}

\item{doctype}{A string declaring the DOCTYPE for the HTML content.}
}
\value{
A HTML document string.
}
\description{
Generate HTML document string with properly declared DOCTYPE.
}
