% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\docType{data}
\name{html5_vars}
\alias{html5_vars}
\title{Environment to set formatted to TRUE/FALSE}
\format{
An environment
}
\usage{
html5_vars
}
\description{
Environment to set formatted to TRUE/FALSE
}
\keyword{datasets}
