% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{html_doc}
\alias{html_doc}
\title{Generate HTML string with properly declared DOCTYPE.}
\usage{
html_doc(..., doctype = "html")
}
\arguments{
\item{...}{A string or strings of HTML element tags.}

\item{doctype}{A string declaring the DOCTYPE for the HTML content.}
}
\value{
A HTML document string.
}
\description{
Generate HTML string with properly declared DOCTYPE.
}
