% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{iframe}
\alias{iframe}
\title{Generate HTML tag for element iframe}
\usage{
iframe(
  accesskey = NULL,
  allow = NULL,
  allowfullscreen = NULL,
  allowpaymentrequest = NULL,
  autocapitalize = NULL,
  class = NULL,
  contenteditable = NULL,
  csp = NULL,
  dir = NULL,
  draggable = NULL,
  height = NULL,
  hidden = NULL,
  id = NULL,
  inputmode = NULL,
  is = NULL,
  itemid = NULL,
  itemprop = NULL,
  itemref = NULL,
  itemscope = NULL,
  itemtype = NULL,
  lang = NULL,
  loading = NULL,
  name = NULL,
  part = NULL,
  referrerpolicy = NULL,
  sandbox = NULL,
  slot = NULL,
  spellcheck = NULL,
  src = NULL,
  srcdoc = NULL,
  style = NULL,
  tabindex = NULL,
  title = NULL,
  width = NULL,
  custom_attr = NULL
)
}
\arguments{
\item{accesskey}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{allow}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{allowfullscreen}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{allowpaymentrequest}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{autocapitalize}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{class}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{contenteditable}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{csp}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{dir}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{draggable}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{height}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{hidden}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{id}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{inputmode}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{is}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemid}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemprop}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemref}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemscope}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemtype}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{lang}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{loading}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{name}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{part}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{referrerpolicy}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{sandbox}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{slot}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{spellcheck}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{src}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{srcdoc}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{style}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{tabindex}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{title}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{width}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{custom_attr}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}
}
\value{
A HTML tag string.
}
\description{
Generate HTML tag for element iframe
}
\examples{
iframe(class = "test")
}
