% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{img}
\alias{img}
\title{Generate HTML tag for element img}
\usage{
img(
  accesskey = NULL,
  alt = NULL,
  autocapitalize = NULL,
  class = NULL,
  contenteditable = NULL,
  crossorigin = NULL,
  decoding = NULL,
  dir = NULL,
  draggable = NULL,
  height = NULL,
  hidden = NULL,
  id = NULL,
  inputmode = NULL,
  is = NULL,
  ismap = NULL,
  itemid = NULL,
  itemprop = NULL,
  itemref = NULL,
  itemscope = NULL,
  itemtype = NULL,
  lang = NULL,
  loading = NULL,
  part = NULL,
  referrerpolicy = NULL,
  sizes = NULL,
  slot = NULL,
  spellcheck = NULL,
  src = NULL,
  srcset = NULL,
  style = NULL,
  tabindex = NULL,
  title = NULL,
  usemap = NULL,
  width = NULL,
  custom_attr = NULL
)
}
\arguments{
\item{accesskey}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{alt}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{autocapitalize}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{class}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{contenteditable}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{crossorigin}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{decoding}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{dir}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{draggable}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{height}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{hidden}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{id}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{inputmode}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{is}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{ismap}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemid}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemprop}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemref}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemscope}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemtype}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{lang}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{loading}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{part}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{referrerpolicy}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{sizes}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{slot}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{spellcheck}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{src}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{srcset}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{style}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{tabindex}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{title}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{usemap}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{width}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{custom_attr}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}
}
\value{
A HTML tag string.
}
\description{
Generate HTML tag for element img
}
\examples{
img(class = "test")
}
