% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{input}
\alias{input}
\title{Generate HTML tag for element input}
\usage{
input(
  accept = NULL,
  accesskey = NULL,
  alt = NULL,
  autocapitalize = NULL,
  autocomplete = NULL,
  autofocus = NULL,
  capture = NULL,
  checked = NULL,
  class = NULL,
  contenteditable = NULL,
  dir = NULL,
  dirname = NULL,
  disabled = NULL,
  draggable = NULL,
  form = NULL,
  formaction = NULL,
  formenctype = NULL,
  formmethod = NULL,
  formnovalidate = NULL,
  formtarget = NULL,
  height = NULL,
  hidden = NULL,
  id = NULL,
  inputmode = NULL,
  is = NULL,
  itemid = NULL,
  itemprop = NULL,
  itemref = NULL,
  itemscope = NULL,
  itemtype = NULL,
  lang = NULL,
  list = NULL,
  max = NULL,
  maxlength = NULL,
  min = NULL,
  minlength = NULL,
  multiple = NULL,
  name = NULL,
  part = NULL,
  pattern = NULL,
  placeholder = NULL,
  readonly = NULL,
  required = NULL,
  size = NULL,
  slot = NULL,
  spellcheck = NULL,
  src = NULL,
  step = NULL,
  style = NULL,
  tabindex = NULL,
  title = NULL,
  type = NULL,
  value = NULL,
  width = NULL,
  custom_attr = NULL
)
}
\arguments{
\item{accept}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{accesskey}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{alt}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{autocapitalize}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{autocomplete}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{autofocus}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{capture}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{checked}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{class}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{contenteditable}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{dir}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{dirname}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{disabled}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{draggable}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{form}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{formaction}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{formenctype}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{formmethod}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{formnovalidate}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{formtarget}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{height}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{hidden}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{id}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{inputmode}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{is}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemid}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemprop}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemref}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemscope}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemtype}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{lang}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{list}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{max}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{maxlength}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{min}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{minlength}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{multiple}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{name}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{part}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{pattern}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{placeholder}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{readonly}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{required}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{size}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{slot}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{spellcheck}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{src}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{step}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{style}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{tabindex}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{title}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{type}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{value}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{width}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{custom_attr}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}
}
\value{
A HTML tag string.
}
\description{
Generate HTML tag for element input
}
\examples{
input(class = "test")
}
