% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{link}
\alias{link}
\title{Generate HTML tag for element link}
\usage{
link(
  accesskey = NULL,
  as = NULL,
  autocapitalize = NULL,
  class = NULL,
  contenteditable = NULL,
  crossorigin = NULL,
  dir = NULL,
  disabled = NULL,
  draggable = NULL,
  hidden = NULL,
  href = NULL,
  hreflang = NULL,
  id = NULL,
  imagesizes = NULL,
  imagesrcset = NULL,
  inputmode = NULL,
  integrity = NULL,
  is = NULL,
  itemid = NULL,
  itemprop = NULL,
  itemref = NULL,
  itemscope = NULL,
  itemtype = NULL,
  lang = NULL,
  media = NULL,
  part = NULL,
  prefetch = NULL,
  referrerpolicy = NULL,
  rel = NULL,
  sizes = NULL,
  slot = NULL,
  spellcheck = NULL,
  style = NULL,
  tabindex = NULL,
  title = NULL,
  type = NULL,
  custom_attr = NULL
)
}
\arguments{
\item{accesskey}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{as}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{autocapitalize}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{class}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{contenteditable}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{crossorigin}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{dir}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{disabled}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{draggable}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{hidden}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{href}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{hreflang}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{id}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{imagesizes}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{imagesrcset}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{inputmode}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{integrity}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{is}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemid}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemprop}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemref}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemscope}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemtype}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{lang}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{media}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{part}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{prefetch}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{referrerpolicy}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{rel}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{sizes}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{slot}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{spellcheck}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{style}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{tabindex}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{title}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{type}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{custom_attr}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}
}
\value{
A HTML tag string.
}
\description{
Generate HTML tag for element link
}
\examples{
link(class = "test")
}
