% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html5.R
\name{textarea}
\alias{textarea}
\title{Generate HTML tag for element textarea}
\usage{
textarea(
  ...,
  accesskey = NULL,
  autocapitalize = NULL,
  autocomplete = NULL,
  autofocus = NULL,
  class = NULL,
  cols = NULL,
  contenteditable = NULL,
  dir = NULL,
  disabled = NULL,
  draggable = NULL,
  form = NULL,
  hidden = NULL,
  id = NULL,
  inputmode = NULL,
  is = NULL,
  itemid = NULL,
  itemprop = NULL,
  itemref = NULL,
  itemscope = NULL,
  itemtype = NULL,
  lang = NULL,
  maxlength = NULL,
  minlength = NULL,
  name = NULL,
  part = NULL,
  placeholder = NULL,
  readonly = NULL,
  required = NULL,
  rows = NULL,
  slot = NULL,
  spellcheck = NULL,
  style = NULL,
  tabindex = NULL,
  title = NULL,
  wrap = NULL,
  custom_attr = NULL
)
}
\arguments{
\item{...}{A string or strings of permitted content for this HTML element tag. The user is responsible for determining what is permissible.}

\item{accesskey}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{autocapitalize}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{autocomplete}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{autofocus}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{class}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{cols}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{contenteditable}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{dir}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{disabled}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{draggable}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{form}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{hidden}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{id}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{inputmode}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{is}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemid}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemprop}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemref}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemscope}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{itemtype}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{lang}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{maxlength}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{minlength}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{name}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{part}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{placeholder}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{readonly}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{required}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{rows}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{slot}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{spellcheck}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{style}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{tabindex}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{title}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{wrap}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}

\item{custom_attr}{A string of permitted content for this HTML attribute. The user is responsible for determining what is permissible.}
}
\value{
A HTML tag string.
}
\description{
Generate HTML tag for element textarea
}
\examples{
textarea(class = "test", "Example")
}
